+++ astergazer/src/main/java/ua/dborisenko/astergazer/service/IChecklistEntryService.java
@@ -4,12 +4,12 @@ import ua.dborisenko.astergazer.exception.ServiceException;
 
 public interface IChecklistEntryService {
 
-    void create(String controlValue, String returnValue, int checklistId) throws ServiceException;
+    void create(String controlValue, String returnValue, Long checklistId) throws ServiceException;
 
-    void update(long id, String controlValue, String returnValue) throws ServiceException;
+    void update(Long id, String controlValue, String returnValue) throws ServiceException;
 
-    void delete(long id) throws ServiceException;
+    void delete(Long id) throws ServiceException;
 
-    String getReturnValue(int checklistId, String controlValue) throws ServiceException;
+    String getReturnValue(Long checklistId, String controlValue) throws ServiceException;
 
 }
