+++ astergazer/src/main/java/ua/dborisenko/astergazer/service/IChecklistService.java
@@ -2,7 +2,7 @@ package ua.dborisenko.astergazer.service;
 
 import java.util.List;
 
-import ua.dborisenko.astergazer.domain.Checklist;
+import ua.dborisenko.astergazer.model.Checklist;
 import ua.dborisenko.astergazer.dto.JsTreeNodeDynamicDto;
 import ua.dborisenko.astergazer.exception.ServiceException;
 
@@ -12,11 +12,11 @@ public interface IChecklistService {
 
     List<JsTreeNodeDynamicDto> getChecklistsTreeDto() throws ServiceException;
 
-    void update(int id, String name) throws ServiceException;
+    void update(Long id, String name) throws ServiceException;
 
-    void delete(int id) throws ServiceException;
+    void delete(Long id) throws ServiceException;
 
-    List<JsTreeNodeDynamicDto> getEntriesTreeDto(int checklistId) throws ServiceException;
+    List<JsTreeNodeDynamicDto> getEntriesTreeDto(Long checklistId) throws ServiceException;
 
     Checklist getByName(String name) throws ServiceException;
