+++ astergazer/src/main/java/ua/dborisenko/astergazer/service/IExtensionService.java
@@ -4,9 +4,9 @@ import ua.dborisenko.astergazer.exception.ServiceException;
 
 public interface IExtensionService {
 
-    void create(String name, int contextId, int scriptId) throws ServiceException;
+    void create(String name, Long contextId, Long scriptId) throws ServiceException;
 
-    void update(int id, int scriptId, String name) throws ServiceException;
+    void update(Long id, Long scriptId, String name) throws ServiceException;
 
-    void delete(int id) throws ServiceException;
+    void delete(Long id) throws ServiceException;
 }
